backend/requirements.txt
flask
gunicorn
flask-cors
google-cloud-firestore
python-dotenv